refactor(view): 구조 단순화
