--- src/main/java/com/sejong/metaservice/core/view/repository/ViewRepository.java
@@ -1,13 +1,0 @@
-package com.sejong.metaservice.core.view.repository;
-
-import com.sejong.metaservice.core.view.domain.View;
-import com.sejong.metaservice.support.common.enums.PostType;
-
-public interface ViewRepository {
-    View save(View view);
-
-    View updateViewCount(View view);
-
-    // findOne 은 포스트 존재 여부 검증 뒤에 수행됩니다.
-    View findOne(PostType postType, Long postId);
-}
