+++ src/main/java/com/sejong/metaservice/domain/view/kafka/ViewEventPublisher.java
@@ -1,4 +1,4 @@
-package com.sejong.metaservice.infrastructure.view.kafka;
+package com.sejong.metaservice.domain.view.kafka;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
