+++ src/main/java/com/sejong/metaservice/domain/view/repository/ViewJPARepository.java
@@ -1,6 +1,6 @@
-package com.sejong.metaservice.infrastructure.view.repository;
+package com.sejong.metaservice.domain.view.repository;
 
-import com.sejong.metaservice.infrastructure.view.entity.ViewEntity;
+import com.sejong.metaservice.domain.view.domain.ViewEntity;
 import com.sejong.metaservice.support.common.enums.PostType;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
